Remove Dynamic Simulation Client
