+++ src/test/java/org/gridsuite/study/server/studycontroller/StudyTestBase.java
@@ -38,8 +38,8 @@ import org.gridsuite.study.server.repository.StudyRepository;
 import org.gridsuite.study.server.repository.rootnetwork.RootNetworkRepository;
 import org.gridsuite.study.server.service.*;
 import org.gridsuite.study.server.service.client.dynamicsecurityanalysis.DynamicSecurityAnalysisClient;
-import org.gridsuite.study.server.service.client.dynamicsimulation.DynamicSimulationClient;
 import org.gridsuite.study.server.service.dynamicmargincalculation.DynamicMarginCalculationRestService;
+import org.gridsuite.study.server.service.dynamicsimulation.DynamicSimulationRestService;
 import org.gridsuite.study.server.service.loadflow.LoadFlowRestService;
 import org.gridsuite.study.server.service.pccmin.PccMinRestService;
 import org.gridsuite.study.server.service.securityanalysis.SecurityAnalysisRestService;
@@ -263,7 +263,7 @@ class StudyTestBase {
     protected ShortCircuitRestService shortCircuitService;
 
     @Autowired
-    protected DynamicSimulationClient dynamicSimulationClient;
+    protected DynamicSimulationRestService dynamicSimulationRestService;
 
     @Autowired
     protected DynamicSecurityAnalysisClient dynamicSecurityAnalysisClient;
@@ -386,7 +386,7 @@ class StudyTestBase {
         voltageInitService.setBaseUri(baseUrl);
         loadflowRestService.setBaseUri(baseUrl);
         shortCircuitService.setBaseUri(baseUrl);
-        dynamicSimulationClient.setBaseUri(baseUrl);
+        dynamicSimulationRestService.setBaseUri(baseUrl);
         dynamicSecurityAnalysisClient.setBaseUri(baseUrl);
         dynamicMarginCalculationRestService.setBaseUri(baseUrl);
         stateEstimationService.setBaseUri(baseUrl);
